Added Menu - Language setting
Default language will be english and if the file is not present it will get the english version.

The font that we currently uses does not have unicode characters in it. Unicode characters will only show if the font contains that unicode. I have provided a sample font, Lusitana, for testing